desktop/requirements.txt
PySide6
qt-material
qtawesome
beautifulsoup4
SQLAlchemy
reportlab
folium
pandas
matplotlib
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
